cpu_input_boost: implement Dynamic SchedTune Boost
Dynamic SchedTune Boost occurs on-the-fly during interactions, such
as touch input. This idea was conceived because of an existing battery
drain issue whereby setting a permanent schedtune boost would cause
boosting to occur even when the phone was idling. For example, if a user
sets the /dev/stune/top-app/schedtune.boost to 10, the device may not
idle at the lowest frequency step, which can lead to higher idle and
active drain rates. Thus it should consume less power to boost only
during interactions.

By default, input boost is 20. The default values can be changed as necessary
in Kconfig.

Setting a value of 0 for any of the boost levels disables boosting for
that level.

History of the original/base implementation by joshuous in the cpu-boost
driver:
  v1 worked by changing the value returned by schedtune_cpu_boost(), but
  did not change the true boost value of the top-app cgroup. In comparison,
  v2 modifies the value in /dev/stune/top-app/schedtune.boost directly.
  v3 reworks the entire Dynamic SchedTune Boost code and improves detection
  of top-app or any other SchedTune group.

This implementation uses v3 of the Dynamic SchedTune Boost framework.

Change-Id: Ief747e847a551a02af38c5bbbf28a37beaf3910a
[celtare21: Remove the ability of the user to change the tunables and remove general stune/ max stune level]
Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: celtare21 <celtare21@gmail.com>